p9: Refactor generic parts of the client into proto.
